+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-Next Release
-============
-* [#11](https://github.com/dblock/guard-rack/pull/11): Use posix-spawn instead of system() - [@viking](https://github.com/viking).
+1.3 (1/15/2013)
+===============
+* [#11](https://github.com/dblock/guard-rack/pull/11): Use [spoon](https://github.com/headius/spoon) instead of `system()`, removes the need for `.pid` files - [@viking](https://github.com/viking), [@dblock](https://github.com/dblock).
 
 1.2.2 (12/19/2012)
 ==================